+++ app/src/main/java/io/bloco/circles/domain/CheckUsernameStatus.kt
@@ -1,20 +1,20 @@
 package io.bloco.circles.domain
 
 import io.bloco.circles.data.network.models.RegistrarName.RegistrarNameStatus
-import io.bloco.circles.data.network.services.RegistrarService
+import io.bloco.circles.data.network.services.LookupService
 import io.bloco.circles.domain.CheckUsernameStatus.UsernameStatus.*
 import retrofit2.HttpException
 import javax.inject.Inject
 
 class CheckUsernameStatus
 @Inject constructor(
-    private val registrarService: RegistrarService
+    private val lookupService: LookupService
 ) {
 
     //TODO: this needs test coverage
     // we use response code and not status itself
     suspend fun isAvailable(username: String): UsernameStatus = try {
-        when(registrarService.lookupName("$username.id.blockstack").status) {
+        when(lookupService.lookupName("$username.id.stx").status) {
             RegistrarNameStatus.Available -> Available
             RegistrarNameStatus.InvalidName -> Invalid
             RegistrarNameStatus.SubmittedSubDomain -> Unavailable