+++ app/src/main/java/io/bloco/circles/domain/NewIdentity.kt
@@ -4,7 +4,8 @@ import io.bloco.circles.data.IdentityModel
 import io.bloco.circles.data.ProfileModel
 import io.bloco.circles.data.interfaces.IdentityRepository
 import kotlinx.coroutines.flow.first
-import org.blockstack.android.sdk.toBtcAddress
+import org.blockstack.android.sdk.extensions.toBtcAddress
+import org.blockstack.android.sdk.extensions.toStxAddress
 import javax.inject.Inject
 
 class NewIdentity
@@ -16,17 +17,26 @@ class NewIdentity
     private val uploadProfile: UploadProfile,
     private val uploadWallet: UploadWallet,
 ) {
-    suspend fun create(username: String?): Result<IdentityModel> = try {
+    suspend fun create(username: String? = null): Result<IdentityModel> = try {
         var identities = identityRepository.observe().first()
-        val keys = identityKeys.new()
-        val btcAddress = keys.toBtcAddress()
 
-        val newIdentity = generateIdentity.generate(btcAddress, username)
+        val keys = identityKeys.new()
+        val btcKeys = keys.toBtcAddress()
+        val stxKeys = keys.toStxAddress(true)
+
+        val newIdentity = generateIdentity.generate(stxKeys, username)
 
         identities = identities + newIdentity
 
         val profile = ProfileModel()
-        username?.also {  registrarProfile.register(username, btcAddress) }
+        username?.also {
+            registrarProfile.register(
+                username,
+                btcKeys,
+                stxKeys
+            )
+        }
+
         uploadProfile.upload(profile, keys)
         uploadWallet.upload(identities)
         identityRepository.set(identities)