+++ app/src/main/java/io/bloco/circles/ui/auth/identities/IdentitiesViewModel.kt
@@ -1,12 +1,6 @@
 package io.bloco.circles.ui.auth.identities
 
-import androidx.hilt.lifecycle.ViewModelInject
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.channels.BroadcastChannel
-import kotlinx.coroutines.flow.*
-import kotlinx.coroutines.launch
-import org.json.JSONObject
+import dagger.hilt.android.lifecycle.HiltViewModel
 import io.bloco.circles.data.AuthRequestModel.AppDetails
 import io.bloco.circles.data.AuthRequestsStore
 import io.bloco.circles.data.AuthResponseModel
@@ -14,20 +8,33 @@ import io.bloco.circles.data.IdentityModel
 import io.bloco.circles.data.interfaces.IdentityRepository
 import io.bloco.circles.domain.GenerateAuthResponse
 import io.bloco.circles.domain.GetAppDetails
+import io.bloco.circles.domain.NewIdentity
 import io.bloco.circles.ui.BaseViewModel
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.channels.BroadcastChannel
+import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.launch
+import org.json.JSONObject
+import timber.log.Timber
+import javax.inject.Inject
 
+@HiltViewModel
 class IdentitiesViewModel
-@ViewModelInject constructor(
+@Inject constructor(
     private val authRequestsStore: AuthRequestsStore,
     generateAuthResponse: GenerateAuthResponse,
     identityRepository: IdentityRepository,
-    getAppDetails: GetAppDetails
+    getAppDetails: GetAppDetails,
+    newIdentity: NewIdentity,
 ) : BaseViewModel() {
 
     // Inputs
     private val identitySelected = BroadcastChannel<IdentityModel>(1)
+    private val createNewIdentity = BroadcastChannel<Unit>(1)
 
     // Outputs
+    private val chooseNewIdentityUsername = BroadcastChannel<Unit>(1)
     private val sendAuthResponse = BroadcastChannel<AuthResponseModel>(1)
     private val identities = BroadcastChannel<List<IdentityModel>>(1)
     private val appDetails = MutableStateFlow<AppDetails?>( null)
@@ -38,7 +45,6 @@ class IdentitiesViewModel
     init {
         identityRepository
             .observe()
-            .take(1)
             .onEach {
                 identities.send(it)
             }
@@ -54,6 +60,21 @@ class IdentitiesViewModel
             }
         }
 
+        createNewIdentity
+            .asFlow()
+            .onEach {
+                authRequestsStore.get()?.let { request ->
+                    if(request.registerSubdomain) {
+                        chooseNewIdentityUsername.send(Unit)
+                    } else {
+                        loading.send(true)
+                        newIdentity.create()
+                        loading.send(false)
+                    }
+                }
+            }
+            .launchIn(ioScope)
+
         identitySelected
             .asFlow()
             .onEach {
@@ -68,18 +89,21 @@ class IdentitiesViewModel
                     )
                 )
             }
-            .catch {
+            .catch { e ->
                 loading.send(false)
                 errors.send(Unit)
                 emitAll(flow { IdentityModel(JSONObject("")) })
+                Timber.e(e)
             }
             .launchIn(ioScope)
     }
 
     // Inputs
     suspend fun identitySelected(identity: IdentityModel) = identitySelected.send(identity)
+    suspend fun createNewIdentity() = createNewIdentity.send(Unit)
 
     // Outputs
+    fun chooseNewIdentityUsername(): Flow<Unit> = chooseNewIdentityUsername.asFlow()
     fun sendAuthResponse(): Flow<AuthResponseModel> = sendAuthResponse.asFlow()
     fun identities(): Flow<List<IdentityModel>> = identities.asFlow()
     fun appDetails(): Flow<AppDetails?> = appDetails.asStateFlow()