+++ app/src/main/java/io/bloco/circles/ui/secret/bottomsheet/ShareKeyBottomSheetFragment.kt
@@ -42,8 +42,8 @@ class ShareKeyBottomSheetFragment : BottomSheetDialogFragment() {
         setStyle(STYLE_NORMAL, R.style.CirclesBottomSheetDialogTheme);
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
         saved
             .clicks()