+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2021-06-09
+### Added
+- Skip username is now based on Auth Request
+### Changed
+- Username lookup service
+
+### Changed
+- Breaking changes with Hilt library
+- Registrar endpoint
+
 ## [0.0.2] - 2021-02-12
 ### Added
 - Ability to Skip Username while creating a new Account/Identity